backend/requirements.txt
mitmproxy==5.2
Django==3.1
djangorestframework==3.11.1
drf-yasg==1.17.1
